refactor(api): move WithMessageContext/GetMessageContext to root package\n\n- Keep MessageContext type in spec\n- Update imports and references across router, consumer, examples, docs\n- All tests and lint pass
