+++ spec/types.go
@@ -86,25 +86,3 @@ type MessageContext struct {
 	Source    string
 	Timestamp string // Application-level timestamp
 }
-
-// messageContextKeyType is an unexported type to avoid context key collisions.
-type messageContextKeyType struct{}
-
-var messageContextKey = messageContextKeyType{}
-
-// WithMessageContext attaches MessageContext to a context and returns the derived context.
-func WithMessageContext(ctx context.Context, mc *MessageContext) context.Context {
-	if mc == nil {
-		return ctx
-	}
-	return context.WithValue(ctx, messageContextKey, mc)
-}
-
-// GetMessageContext extracts the MessageContext from a context.
-func GetMessageContext(ctx context.Context) (*MessageContext, bool) {
-	if ctx == nil {
-		return nil, false
-	}
-	mc, ok := ctx.Value(messageContextKey).(*MessageContext)
-	return mc, ok
-}
